GardeningBot cancel mechanism improved: IPC mechanism (sockets)
git-svn-id: http://halodev.ontoprise.com/svn/HaloSMWExtension/trunk@844 f80b7aa0-dfcc-4e65-a0d9-74108174cbc8

+++ extensions/SMWHalo/includes/SMW_GlobalFunctionsForSpecials.php
@@ -97,9 +97,13 @@ function smwfCancelGardeningBot($taskid) {
 	if (!GardeningBot::isUserAllowed(array(SMW_GARD_SYSOPS, SMW_GARD_GARDENERS))) {
 	 	return; // only sysops and gardeners may cancel a bot.
 	}
-	$processID = GardeningBot::getProcessID($taskid);
-	if ($processID != NULL) {
-		GardeningBot::killProcess($processID);
+	// send term signal to bot
+	if (GardeningBot::abortBot($taskid) !== true) {
+		// if bot does not react: kill process
+		$processID = GardeningBot::getProcessID($taskid);
+		if ($processID != NULL) {
+			GardeningBot::killProcess($processID);
+		}
 	}
 	SMWGardening::getGardeningLogAccess()->removeGardeningTask($taskid);
 	return SMWGardening::getGardeningLogTable();
